Refactored complete app into child apps: consent, conformance
