+++ cmapi/urls.py
@@ -14,10 +14,10 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
-from django.urls import path, re_path, include
-
+from django.urls import include, path, re_path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    re_path(r'^', include('consent.urls')),
+    re_path(r'^', include('consent.urls', namespace='consent')),
+    re_path(r'^', include('conformance.urls', namespace='conformance')),
 ]
