+++ conformance/templates/manageresources.html
@@ -1,15 +1,8 @@
 {% extends 'base.html' %}
 
-{% block body %}
+{% block manageresources_active %}active{% endblock %}
 
-  <nav class="navbar navbar-default">
-    <ul class="nav navbar-nav">
-      <li><a href="{% url 'consent:index'%}">Home</a></li>
-      <li><a href="{% url 'consent:consentapproval'%}">Consent Approval</a></li>
-      <li class="active"><a href="{% url 'consent:manageresources'%}">Manage Resources</a></li>
-      <li><a href="{% url 'consent:capabilitystatement'%}">Capability Statement</a></li>
-    </ul>
-  </nav>
+{% block body %}
 
   <div class="page-header">
     <h1>Welcome to Resource Manager Portal</h1>
