+++ consent/templates/index.html
@@ -173,15 +173,6 @@
     });
   </script>
 
-  <nav class="navbar navbar-default">
-    <ul class="nav navbar-nav">
-      <li class="active"><a href="{% url 'consent:index'%}">Home</a></li>
-      <li><a href="{% url 'consent:consentapproval'%}">Consent Approval</a></li>
-      <li><a href="{% url 'consent:manageresources'%}">Manage Resources</a></li>
-      <li><a href="{% url 'consent:capabilitystatement'%}">Capability Statement</a></li>
-    </ul>
-  </nav>
-
   <div class="page-header">
     <h1>Welcome to Caregiver Portal</h1>
   </div>
