+++ consent/urls.py
@@ -8,6 +8,4 @@ urlpatterns = [
     re_path(r'^$', views.index, name='index'),
     re_path(r'^consentapproval/$', views.consentapproval, name='consentapproval'),
     re_path(r'^displaydata/(?P<requestid>[0-9]+)/$', views.displaydata, name='displaydata'),
-    re_path(r'^manageresources/$', views.upload_resource, name='manageresources'),
-    re_path(r'^capabilitystatement/$', views.get_capability_statement, name='capabilitystatement'),
 ]
