+++ templates/base.html
@@ -31,6 +31,12 @@
           <span class="icon-bar"></span>
         </button>
         <a class="navbar-brand" href="{% url 'consent:index'%}">Consent & Conformance POC</a>
+        <ul class="nav navbar-nav">
+          <!-- <li class="active"><a href="{% url 'consent:index'%}">Home</a></li> -->
+          <li class="{% block consentapproval_active %}{% endblock %}"><a href="{% url 'consent:consentapproval'%}">Consent Approval</a></li>
+          <li class="{% block manageresources_active %}{% endblock %}"><a href="{% url 'conformance:manageresources'%}">Manage Resources</a></li>
+          <li class="{% block capabilitystatement_active %}{% endblock %}"><a href="{% url 'conformance:capabilitystatement'%}">Capability Statement</a></li>
+        </ul>
       </div>
     </div>
   </nav>
